Update README.md
Organization "We Remember August" was created by immigrants living in the United States.
The organization's mission is first and foremost to help families of military personnel killed in War August 2008 against Russia.

These families receive financial assistance from the government but we believe that if fans of all patriots, regardless of their whereabouts, personal actions can in memory of fallen heroes and their families for their support, respect and gratitude to express.

Anyone has the right to join the organization. Donations are voluntary. The amount of money collected in the form of a fixed monthly donation call distribution process much easier.

The sums collected on a monthly basis, for one year, will be distributed to several families.

The organization provides financial support for the collection, distribution and delivery of a foster family for complete transparency and accountability.